refactored code to export widgets in its own files
